requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==1.0.6
Werkzeug==0.10.4
PyGithub==1.25.2
Flask-OAuthlib==0.9.1

